requirements.txt
gymnasium==0.29.1
matplotlib==3.8.2
numpy==1.21.5
pandas
burnysc2
stable_baselines3
